backend/requirements.txt
fastapi
uvicorn
qdrant-client
transformers
torch
pillow
sentencepiece
hf_xet
python-multipart